+++ tests/assets/c_source_2/file_with_private_members.c
@@ -25,7 +25,7 @@ static int my_var;
 float a_public_var;
 
 /**
- * Enumerations in c files are inherently private
+ * Enums in c files are inherently private
  */
 enum private_enum {
     ENUM_1